refactor(app): split storage settings editor
